speech_to_text/requirements.txt
streamlit>=1.28.0
openai-whisper>=20250625
pyperclip>=1.8.2
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
tqdm>=4.64.0
ffmpeg-python>=0.2.0
pathlib2>=2.3.7
